synthesis-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-firestore==2.17.0

